refactor(diagnostics): creating a separate file for diagnostics related utilities
