+++ src/arkreactor/Compiler/Welder.cpp
@@ -6,7 +6,8 @@
 #include <Ark/Compiler/Macros/Processor.hpp>
 #include <Ark/Compiler/NameResolution/NameResolutionPass.hpp>
 #include <Ark/Utils/Files.hpp>
-#include <Ark/Exceptions.hpp>
+#include <Ark/Error/Exceptions.hpp>
+#include <Ark/Error/Diagnostics.hpp>
 
 #include <sstream>
 #include <fmt/ostream.h>
