+++ src/arkreactor/VM/VM.cpp
@@ -8,6 +8,7 @@
 
 #include <Ark/Utils/Files.hpp>
 #include <Ark/Utils/Utils.hpp>
+#include <Ark/Error/Diagnostics.hpp>
 #include <Ark/TypeChecker.hpp>
 #include <Ark/VM/ModuleMapping.hpp>
 #include <Ark/Compiler/Instructions.hpp>
